+++ src/app/features/currencies/components/currency-creation-form/currency-creation-form.ts
@@ -4,12 +4,12 @@ import { DialogModule } from 'primeng/dialog';
 import { InputTextModule } from 'primeng/inputtext';
 import { MessageModule } from 'primeng/message';
 import { ButtonModule } from 'primeng/button';
-import { CurrencyCreationModalService } from '@features/currencies/services/currency-creation-modal.service';
+import { CurrencyCreationFormService } from '@features/currencies/services/currency-creation-form.service';
 import { RequiredStar } from '@toolkit/components/required-star/required-star';
 
 @Component({
-	selector: 'gb-currencies-currency-creation-modal',
-	templateUrl: './currency-creation-modal.html',
+	selector: 'gb-currencies-currency-creation-form',
+	templateUrl: './currency-creation-form.html',
 	imports: [
 		ReactiveFormsModule,
 		DialogModule,
@@ -19,6 +19,6 @@ import { RequiredStar } from '@toolkit/components/required-star/required-star';
 		RequiredStar,
 	],
 })
-export class CurrencyCreationModal {
-	readonly modal = inject(CurrencyCreationModalService);
+export class CurrencyCreationForm {
+	readonly formService = inject(CurrencyCreationFormService);
 }